Split tools_issue_lifecycle.py into tools_issue_headless.py and tools_issue_labels.py
- Created tools_issue_headless.py with prepare_issue and enrich_issues (headless session tools)
- Created tools_issue_labels.py with claim_issue and release_issue (GitHub label management)
- Deleted tools_issue_lifecycle.py
- Updated server/__init__.py imports
- Updated all test imports and arch tests
- Updated CLAUDE.md and tool-access.md documentation
