+++ src/autoskillit/server/__init__.py
@@ -103,7 +103,10 @@ from autoskillit.server.tools import (  # noqa: E402, F401
     tools_issue_composite as _tools_issue_composite,
 )
 from autoskillit.server.tools import (  # noqa: E402, F401
-    tools_issue_lifecycle as _tools_issue_lifecycle,
+    tools_issue_headless as _tools_issue_headless,
+)
+from autoskillit.server.tools import (  # noqa: E402, F401
+    tools_issue_labels as _tools_issue_labels,
 )
 from autoskillit.server.tools import (  # noqa: E402, F401
     tools_kitchen as _tools_kitchen,
