+++ src/autoskillit/server/tools/CLAUDE.md
@@ -20,7 +20,8 @@ MCP `@mcp.tool()` handlers registered on import (16 tool modules).
 | `tools_fleet_dispatch.py` | `dispatch_food_truck`, `record_gate_dispatch` |
 | `tools_git.py` | `merge_worktree`, `classify_fix`, `create_unique_branch`, `create_and_publish_branch`, `check_pr_mergeable` |
 | `tools_github.py` | `fetch_github_issue`, `get_issue_title`, `report_bug` |
-| `tools_issue_lifecycle.py` | `prepare_issue`, `enrich_issues`, `claim_issue`, `release_issue` |
+| `tools_issue_headless.py` | `prepare_issue`, `enrich_issues` (headless session tools) |
+| `tools_issue_labels.py` | `claim_issue`, `release_issue` (GitHub label management) |
 | `tools_issue_composite.py` | `claim_and_resolve_issue` |
 | `tools_pr_ops.py` | `get_pr_reviews`, `bulk_close_issues` |
 | `tools_recipe.py` | `load_recipe`, `list_recipes`, `validate_recipe`, `migrate_recipe` |
