+++ tests/arch/test_layer_enforcement.py
@@ -273,7 +273,8 @@ def test_ungated_tools_do_not_call_require_enabled() -> None:
         tools_execution,
         tools_git,
         tools_github,
-        tools_issue_lifecycle,
+        tools_issue_headless,
+        tools_issue_labels,
         tools_pr_ops,
         tools_recipe,
         tools_status,
@@ -288,7 +289,8 @@ def test_ungated_tools_do_not_call_require_enabled() -> None:
         tools_recipe,
         tools_status,
         tools_github,
-        tools_issue_lifecycle,
+        tools_issue_headless,
+        tools_issue_labels,
         tools_pr_ops,
     ]
