+++ tests/recipe/test_rules_tools.py
@@ -344,7 +344,8 @@ _SERVER_TOOL_MODULES = [
     "autoskillit.server.tools.tools_recipe",
     "autoskillit.server.tools.tools_status",
     "autoskillit.server.tools.tools_github",
-    "autoskillit.server.tools.tools_issue_lifecycle",
+    "autoskillit.server.tools.tools_issue_headless",
+    "autoskillit.server.tools.tools_issue_labels",
     "autoskillit.server.tools.tools_pr_ops",
     "autoskillit.server.tools.tools_workspace",
     "autoskillit.server.tools.tools_agents",
