+++ tests/server/test_release_issue_fail_label.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from autoskillit.server.tools.tools_issue_lifecycle import claim_issue, release_issue
+from autoskillit.server.tools.tools_issue_labels import claim_issue, release_issue
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
