+++ tests/server/test_tools_integrations_release.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from autoskillit.server.tools.tools_issue_lifecycle import release_issue
+from autoskillit.server.tools.tools_issue_labels import release_issue
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
